+++ core/services/pipeline/task_helpers_test.go
@@ -1,0 +1,5 @@
+package pipeline
+
+func (t *ETHTxTask) SetGasLimitGwei(v uint32) {
+	t.gasLimitGwei = &v
+}