+++ core/web/presenters/job.go
@@ -396,6 +396,7 @@ type JobResource struct {
 	Name                   string                  `json:"name"`
 	Type                   JobSpecType             `json:"type"`
 	SchemaVersion          uint32                  `json:"schemaVersion"`
+	GasLimitGwei           *uint32                 `json:"gasLimitGwei"`
 	MaxTaskDuration        models.Interval         `json:"maxTaskDuration"`
 	ExternalJobID          uuid.UUID               `json:"externalJobID"`
 	DirectRequestSpec      *DirectRequestSpec      `json:"directRequestSpec"`
@@ -419,6 +420,7 @@ func NewJobResource(j job.Job) *JobResource {
 		Name:            j.Name.ValueOrZero(),
 		Type:            JobSpecType(j.Type),
 		SchemaVersion:   j.SchemaVersion,
+		GasLimitGwei:    j.GasLimitGwei,
 		MaxTaskDuration: j.MaxTaskDuration,
 		PipelineSpec:    NewPipelineSpec(j.PipelineSpec),
 		ExternalJobID:   j.ExternalJobID,