+++ core/web/presenters/job_test.go
@@ -45,6 +45,8 @@ func TestJob(t *testing.T) {
 	v2CoordAddress, err := ethkey.NewEIP55Address("0x2C409DD6D4eBDdA190B5174Cc19616DD13884262")
 	require.NoError(t, err)
 
+	var gasLimitGwei uint32 = 1000
+
 	testCases := []struct {
 		name string
 		job  job.Job
@@ -53,7 +55,8 @@ func TestJob(t *testing.T) {
 		{
 			name: "direct request spec",
 			job: job.Job{
-				ID: 1,
+				ID:           1,
+				GasLimitGwei: &gasLimitGwei,
 				DirectRequestSpec: &job.DirectRequestSpec{
 					ContractAddress: contractAddress,
 					CreatedAt:       timestamp,
@@ -99,6 +102,7 @@ func TestJob(t *testing.T) {
 						"offChainReportingOracleSpec": null,
 						"offChainReporting2OracleSpec": null,
 						"fluxMonitorSpec": null,
+						"gasLimitGwei": 1000,
 						"keeperSpec": null,
                         "cronSpec": null,
                         "vrfSpec": null,
@@ -168,6 +172,7 @@ func TestJob(t *testing.T) {
 							"updatedAt":"2000-01-01T00:00:00Z",
 							"evmChainID": "42"
 						},
+						"gasLimitGwei": null,
 						"offChainReportingOracleSpec": null,
 						"offChainReporting2OracleSpec": null,
 						"directRequestSpec": null,
@@ -252,6 +257,7 @@ func TestJob(t *testing.T) {
 						},
 						"offChainReporting2OracleSpec": null,
 						"fluxMonitorSpec": null,
+						"gasLimitGwei": null,
 						"directRequestSpec": null,
 						"keeperSpec": null,
                         "cronSpec": null,
@@ -309,6 +315,7 @@ func TestJob(t *testing.T) {
 							"evmChainID": "42"
 						},
 						"fluxMonitorSpec": null,
+						"gasLimitGwei": null,
 						"directRequestSpec": null,
 						"cronSpec": null,
 						"webhookSpec": null,
@@ -364,6 +371,7 @@ func TestJob(t *testing.T) {
                             "updatedAt":"2000-01-01T00:00:00Z"
                         },
                         "fluxMonitorSpec": null,
+						"gasLimitGwei": null,
                         "directRequestSpec": null,
                         "keeperSpec": null,
                         "offChainReportingOracleSpec": null,
@@ -416,6 +424,7 @@ func TestJob(t *testing.T) {
 							"updatedAt":"2000-01-01T00:00:00Z"
 						},
 						"fluxMonitorSpec": null,
+						"gasLimitGwei": null,
 						"directRequestSpec": null,
 						"keeperSpec": null,
 						"cronSpec": null,
@@ -467,6 +476,7 @@ func TestJob(t *testing.T) {
 						"externalJobID": "0eec7e1d-d0d2-476c-a1a8-72dfb6633f46",
 						"directRequestSpec": null,
 						"fluxMonitorSpec": null,
+						"gasLimitGwei": null,
 						"cronSpec": null,
 						"offChainReportingOracleSpec": null,
 						"offChainReporting2OracleSpec": null,
@@ -529,6 +539,7 @@ func TestJob(t *testing.T) {
 						"externalJobID": "0eec7e1d-d0d2-476c-a1a8-72dfb6633f46",
 						"directRequestSpec": null,
 						"fluxMonitorSpec": null,
+						"gasLimitGwei": null,
 						"cronSpec": null,
 						"offChainReportingOracleSpec": null,
 						"offChainReporting2OracleSpec": null,
@@ -612,6 +623,7 @@ func TestJob(t *testing.T) {
 							"evmChainID": "42"
 						},
 						"fluxMonitorSpec": null,
+						"gasLimitGwei": null,
 						"directRequestSpec": null,
 						"cronSpec": null,
 						"webhookSpec": null,