+++ docs/CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added 
+- Added job spec attribute `gasLimitGwei` which overrides `ETH_GAS_LIMIT_DEFAULT` value.
+
 ## [1.5.0] - 2022-06-21
 
 ### Changed